backend/cybersec/requirements.txt
Django==5.0.3
python-dotenv==1.0.1
Requests==2.31.0
django-cors-headers==4.3.1